'on_posting' and 'on_posting_<resource>' event hooks.
'on_posting' and 'on_posting_<resource>' events are raised when
documents are about to be stored. Among other things this allows
callback functions to arbitrarily update the documents being inserted.

'on_posting' is raised on every successful POST while 'on_posting_<resource>'
is only raised when <resource> is being updated. In both circumstances the
event will be raised only if at least one document passed validation and is
going to be inserted.