Optional custom payload for methods.post. Addresses #74.
When calling post() from your own code, you can provide an alternative
payload. This can be useful, for example, when you have a callback
function hooked to a certain endpoint, and want to perform additional
post() calls from there.

Please be advised that in order to successfully use this option,
a request context must be available.

See https://github.com/nicolaiarocci/eve/issues/74 for a discussion and
a typical use case.