Filters enabled by default.
Since I'm not aware of direct attack vectors on mongo 'where' clause,
except for JS operators which are blacklisted by default already,
I prefer leaving filters enabled as it will make API implementation
easier on the API maintainer (not having to explictly whitelist every
single filter, for every resource).

In the documentation I'm going to suggest that if scraping or preventing
DB DoS attacks are a concern, then disabling all filters and then
whitelisting valid ones is the way to go.

NOTE: not really sure on this one, we might fallback to the original "filters
disabled by default" proposed by Bryan Cattle.