Add JSONP support
With this feature you're able to add a JSONP_ARGUMENT to your config.
When you request a resource you can add this argument and the response
will be wrapped in a JavaScript function call. The obviously only works
with JSON. If XML is requested the setting is ignored.

Example:
request: curl -i http://localhost:5000/?callback=JSONP_CALLBACK
response: JSONP_CALLBACK( # json response here )