Auth field not set if resource level auth is set.
Fixes #231.

When authentication is set globally, POSTing a document with the correct
credentials will set the the auth field on the new document before
inserting into mongo. If the authentication object is set on the
resource configuration, the auth_field is not set on the newly inserted
document. This causes a problem if I later want to do a GET, PUT, or
PATCH request on an object I've inserted that has resource level
authentication.