Fixes # 243.
GET with If-Modified-Since on resource endpoint returns incorrect 304 if
resource is empty. Not terribly happy with this 'fix' as it can cause
noticeable performance issues if database is huge and/or not optimzed.
This is a case where in order to fix a very rare occurence we're going
to cause an impact on almost all IMS requests. Putting it up for review,
but it should probably be reverted and #243 marked as Won't Fix (or find
a better alternative).