Added on_pre_GET_<resource>_lookup event.
This feature allows to programmatically alter the lookup parameter passed to the data
layer that will be used to find documents. The contraint returned by a lookup event
is not restricted any `allowed_filters` setting since it is considered trusted. This
allows the API to be restricted in terms of what filters clients can supply, while
still providing full flexibility to developers.

This feature is incomplete. If discussions are positive, it will be completed and
tests added.