Add support for two layers of document embedding
This feature is incomplete (docs and tests are yet to be written). It's
a preview to make sure the approach is sane.

This feature supports document embedding with dotted notation one layer
deep as long as the second layer is a list of dicts. E.g. users.friends
would be supported provided the schema declares 'users' to be a 'list'
type and 'friends' to be a dict.