Add tox support
I've started with the following versions: 2.6, 2.7, 3.3, 3.4 and pypy.
Remove the ones you don't want.

The flake8 step is added as a tox test environment.

There was a problem with the config.py test. Where the custom name was
tested. I've found some problems in pypy and py34 with that test because
the name of the module wasn't an existing one, so after playing with
multiple options, the easiest patch was to use the name of an existing
module.