HATEOAS links as business unit values.
HATOEAS link for the item and resource espresses the business unit value
even when regexes are configured for the resource endpoint. This applies
to resource and item endpoints. Home links are still showing
the regexes (as they are not computable based on the current request).
To address these I plan to add a filter/parser in the near future.