Reduce chances of failure on sort_disabled test
Previously we were iterating the whole resultset which might cause
occasional failures as mongodb natural order is not granted to be the
insertion order. In theory this test can still fail on occasion,
although it is highly unlikely.

For the record this test was costantly failing on a MacBook Air where
mongodb and the whole python/eve toolchain were just installed. Odd
since nobody reported this before, but still.