HATEOAS links are now all relative to API root.
"Could we make EVE run with the assumption of relative paths instead of
absolute? The current infrastructure seems like it is thinking that
flask is the end web server that is used in production and in fact most
people will be using this with Nginx or Apache. Having to keep multiple
versions of the settings.py file to support different host names seems
unneeded and redundant."

Addresses #401 #398.