Fix validation of read-only fields inside dicts.
Issue #474 : As the default values are resolved before validation,
the validation method of readonly fields needs to know about an
 eventual default value of the field in order to skip validation if
the value being validated matches the default.

The default value was being fetched from the schema definition, which
raised a KeyError when the read-only field was not in the first level of
the document, but inside a dict.

This is fixed by getting the default value from the schema attribute
instead.

closes #474