Adapt test auth classes and write tests to demonstrate bug with changing config
The bug in question is that when both `allowed_roles` (can be empty list) and
`allowed_read_roles` or `allowed_item_roles` are specified, the `allowed_roles`
list in the app's config will grow over time, expanding with the items in
`allowed_read_roles` or `allowed_write_roles` at every request that considers
the roles.

This is also true for:

 - ALLOWED_ROLES (changes if ALLOWED_READ_ROLES or ALLOWED_WRITE_ROLES are
   specified)
 - allowed_item_roles (changes if allowed_item_read_roles or
   allowed_item_write_roles are specified)

In order to demonstrate this, the test auth classes `ValidBasicAuth` are changed
so that their `check_auth` method returns `True` if 'admin' is at least one of
the items in the given `allowed_roles`. Previously, they would only accept if
'admin' was the only element of the list.

In addition, three tests are added, specifically testing that the allowed roles
in the app's config do not change between requests.