Fix with expanding roles lists if read/write roles counterpart is specified
Previously, the `allowed_roles`, `allowed_item_roles` and config `ALLOWED_ROLES`
were fetched as a reference and expanded with items from their read/write
sub-specifications, that is:

 - `allowed_read_roles` and `allowed_write_roles` for `allowed_roles`
 - `allowed_item_read_roles` and `allowed_item_write_roles` for
   `allowed_item_roles`
 - `ALLOWED_READ_ROLES` and `ALLOWED_WRITE_ROLES` for `ALLOWED_ROLES`

This becomes a problem for two reasons:

 - **SECURITY:** the list will potentially expand with unintended roles.
   Read-only roles will be copied over to 'allowed roles', allowing access also
   for writing in the next request.
 - **MEMORY:** the list grows without bounds, forever using more and more
   memory.

The problem could be solved in config by specifying roles in tuples, which are
immutable, instead of lists, which can expand. However, this could be confusing
and is not documented. This commit solves the issue in code, by copying the
lists with `list(list_to_be_copied)` instead of using the reference directly.