User Restricted Access works with HMACAuth classes.
When eve.io.auth.HMACAuth subclasses were used for client
authentication, the User Restricted Resource Access was not working.

This was due to a (hopefully) rendundant check on request.authorization,
which was performed to verify that URRA was active. Flask's
request.authorization will be None whne HMAC is used (since the
Authorization header format is different from the one used with Basic
Auth.)