Revert #561 so it works with a pre-#556 codebase.
'Hardening of concurrency checks' (#561) depends updates introduced with

We are preparing for v0.5.2, which is a 'bugfix' release. #556 adds
a new feature and it won't go into this release, so we need to fix #561
code in order to get it to work with a pre-#556 codebase.

This is supposed to be an orphaned commit, it won't exist outside of
v0.5.2 branch.