Added 'MEDIA_CUSTOM_URL' setting
Allows media urls to use a custom URL as a base pattern.
When 'RETURN_MEDIA_AS_URL' is set to True and 'MEDIA_CUSTOM_URL' is not None, any media files
in a collection will return a concatenation of this setting plus the file ID.

This allows custom 'MediaStorage' to be much more flexible.