embedded documents following depth level
embedded_fields may point on a field that come from another embedded
document. For example, ['a.b.c', 'a.b', 'a']

The strategy is to sort out the different embedded_fields depending on
their number of '.'. This way, the ones with the greatest depth level
come last.

Combined with https://github.com/RedTurtle/eve-sqlalchemy/pull/47 it's
now possible to do include multi-level embedded resource with Eve-SQLAlchemy.