VALIDATIN_ERROR_AS_LIST. Defaults to False.
If True even single field errors will be returned in a list. By
default single field errors are returned as strings while multiple field
errors are bundled in a list. If you want to standardize the field
errors output, set this setting to True and you will always get
a list of field issues. Defaults to False.

Closes #536.