Fix filter validation
The `eve.utils.validate_filters` function aims to validate a MongoDB
query selector. To do so, for each key-value pair of the query, it
matches the key against a set of allowed fields and operators and it
recursively validates the associated value if it's a dict or a list,
under the assumption that it's respectively a sub-query or a list of
sub-queries.

While this assumption seems to be legitimate if the value is a dict,
it's not if the value is a list. For instance, the '$in' operator
expects a list of field values, not a list of sub-queries.

In fact, it's legitimate to assume that a value of the list data-type is
a list of sub-queries only if the corresponding key is one the following
operators: '$or', '$and' and '$not'.

Thus, if we encounter a value of the list data-type, we should try to
recursively validate it only if the corresponding key is either '$or',
'$and' or '$not'.