Exclude ID_FIELD from versioned fields
When versioning is enabled on a resource with a custom ID_FIELD,
versioning documents will inherit their ID from the versioned document,
making any update of the document result in a `DuplicateKeyError`.

Since there is no point in versioning an ID, let's just exclude it from
the list of versioned fields.