PUT attempts to create a document if it does not exist.
The URL endpoint will be used as ``ID_FIELD`` value (if ``ID_FIELD`` is
included with the payload, it will be ignored). Normal validation rules
apply. The response will be a ``201 Created`` on successful creation.
Response payload will be identical the one you would get by performing
a single document POST to the resource endpoint.

PUT does not return ``404`` if a document URL does not exist. It will
attempt to create the document instead.

Closes #634.