Support ENFORCE_IF_MATCH flag
If `ENFORCE_IF_MATCH` is disabled and `IF_MATCH` is enabled, then
the client can choose whether to provide the `If-Match` header in
its request. When `If-Match` is provided, then the request is treated as
conditional, meaning the ETag must match. When `If-Match` is absent,
then the ETag check is not enforced.

The `ENFORCE_IF_MATCH` flag defaults to `True`

Closes #657