Allow posting lists by using the same key multiple times
HTTP allows multiple values with the same key in
application/x-www-form-urlencoded and multipart/form-data content types. We can
use this feature to send a list using for example an HTML form. This is very
useful to send a list of files.
This commit adds a config setting AUTO_COLLAPSE_MULTI_KEYS to enable automatic
conversion of data, which has the same key, to a list of values.

See issue #932