Implement tests for CORS regex settings
A test was added to address the issue that regexes are not fully matched and
the server may let origins that contain extra characters through.

Add some tests that introduce a new setting X_DOMAINS_RE which defines a list
of regexes to match for when deciding which origins are allowed for CORS.
X_DOMAINS should not contain regexes anymore and will only match if the urls in
the list are fully equivalent to the origin.

Implement a test to check that invalid regexes are ignored