Make merging of nested documents optional (enabled by default)
Rationale
---------

In the fix for #519, merging of nested documents on `PATCH` was introduced,
essentially a `original.update(updates)` is performed.
So far, the only way to avoid this is to use `PUT` (see #712), however,
this is not always desirable.

We have a use case where we only want to update a nested field of a
document, therefore the appropriate HTML verb is `PATCH` since we do not want
to replace the whole document.
In this nested field, we need to remove a key, which, due to the merging,
is currently impossible.

Changes
-------

- Introduce a new setting, `MERGE_NESTED_DOCUMENTS`, to control this behaviour
  globally or per-resource.
  Per default, this is set to `True`, so that nothing changes compares to the
  current version of Eve except when the configuration is changed explicitly.
- Add a test
- Add documentation