NORMALIZE_DOTTED_FIELDS/normalize_dotted_fields
If True, dotted fields are parsed and processed as subdocument fields.
If False, dotted fields are left unparsed and unprocessed and the
payload is passed to the underlying data-layer as-is. Please note that
with the default Mongo layer, setting this to False will result in an
error. Defaults to True.

Closes #1173