Fix HEADER_TOTAL_COUNT on emtpy resources
the count returned by `app.find` might be None if the count is disabled.
Later in get_internal, the HEADER_TOTAL_COUNT is prepared if the count
is enabled.

This commit fixes a bug that filtered out the HEADER_TOTAL_COUNT when
the resource is empty and the count is 0, replacing a truthy
condition by an explicit `is not None` condition.