Fix the mongo_options work by passing the options.
This is done by passing the options directly into the find_one method.
The work doesn't cover adding arguments to find_one_raw or find since
they aren't needed to fix the bug.

I explored adding a fluid type interface for with_optios to the eve
PyMongos/Mongo/DataLayer classes but it was really clunky and I'd be
worried that such a fundamental change wouldn't be appropriate here.